Update bot.py
Still a lot of code for testing purposes that will ultimately be discarded.

Worked out how to retrieve channel ids,  handle reactions, etc.

Still needing to decided how to best store data that will need to persist without relying on a list of member objects, as that would require 100% uptime for the bot.